feat(common): Gutting stream util
