+++ module/base/src/util.ts
@@ -1,5 +1,7 @@
+import fs from 'node:fs/promises';
 import crypto from 'node:crypto';
 import timers from 'node:timers/promises';
+import { Readable } from 'node:stream';
 
 import { ManifestFileUtil } from '@travetto/manifest';
 
@@ -123,5 +125,43 @@ export class Util {
    */
   static queueMacroTask(): Promise<void> {
     return timers.setImmediate(undefined);
+  }
+
+  /**
+   * Fetch bytes from a url
+   */
+  static async fetchBytes(url: string, byteLimit: number = -1): Promise<Buffer> {
+    const str = await fetch(url, {
+      headers: (byteLimit > 0) ? {
+        Range: `0-${byteLimit - 1}`
+      } : {}
+    });
+
+    if (!str.ok) {
+      throw new Error('Invalid url for hashing');
+    }
+
+    let count = 0;
+    const buffer: Buffer[] = [];
+
+    for await (const chunk of Readable.fromWeb(str.body!)) {
+      if (Buffer.isBuffer(chunk)) {
+        buffer.push(chunk);
+        count += chunk.length;
+      } else if (typeof chunk === 'string') {
+        buffer.push(Buffer.from(chunk));
+        count += chunk.length;
+      }
+
+      if (count > byteLimit && byteLimit > 0) {
+        break;
+      }
+    }
+
+    try {
+      await str.body?.cancel();
+    } catch { }
+
+    return Buffer.concat(buffer, byteLimit <= 0 ? undefined : byteLimit);
   }
 }
